All the DocViewer elements now live in the docviewer file
