+++ nw/gui/__init__.py
@@ -2,11 +2,11 @@
 
 from nw.gui.about import GuiAbout
 from nw.gui.build import GuiBuildNovel
-from nw.gui.docbars import GuiDocTitleBar, GuiSearchBar, GuiDocViewDetails
+from nw.gui.docbars import GuiDocTitleBar, GuiSearchBar
 from nw.gui.doceditor import GuiDocEditor
 from nw.gui.docmerge import GuiDocMerge
 from nw.gui.docsplit import GuiDocSplit
-from nw.gui.docviewer import GuiDocViewer
+from nw.gui.docviewer import GuiDocViewer, GuiDocViewDetails
 from nw.gui.itemdetails import GuiItemDetails
 from nw.gui.itemeditor import GuiItemEditor
 from nw.gui.mainmenu import GuiMainMenu
@@ -25,11 +25,11 @@ __all__ = [
     "GuiBuildNovel",
     "GuiDocTitleBar",
     "GuiSearchBar",
-    "GuiDocViewDetails",
     "GuiDocEditor",
     "GuiDocMerge",
     "GuiDocSplit",
     "GuiDocViewer",
+    "GuiDocViewDetails",
     "GuiItemDetails",
     "GuiItemEditor",
     "GuiMainMenu",
