+++ nw/gui/docbars.py
@@ -7,9 +7,7 @@
 
  File History:
  Created: 2019-09-29 [0.2.1] GuiSearchBar
- Created: 2019-10-31 [0.3.2] GuiDocViewDetails
  Created: 2020-04-25 [0.4.5] GuiDocTitleBar
- Created: 2020-06-09 [0.8.0] GuiDocViewFooter
 
  This file is a part of novelWriter
  Copyright 2020, Veronica Berglyd Olsen
@@ -353,198 +351,3 @@ class GuiDocTitleBar(QWidget):
         return
 
 # END Class GuiDocTitleBar
-
-class GuiDocViewFooter(QWidget):
-
-    def __init__(self, theParent):
-        QWidget.__init__(self, theParent)
-
-        logger.debug("Initialising GuiDocViewFooter ...")
-
-        self.mainConf  = nw.CONFIG
-        self.theParent = theParent
-        self.theTheme  = theParent.theTheme
-        self.optState  = theParent.theProject.optState
-        self.viewMeta  = theParent.theParent.viewMeta
-        self.theHandle = None
-
-        # Make a QPalette that matches the Syntax Theme
-        self.thePalette = QPalette()
-        self.thePalette.setColor(QPalette.Window, QColor(*self.theTheme.colBack))
-        self.thePalette.setColor(QPalette.Text, QColor(*self.theTheme.colText))
-
-        fPx = int(0.9*self.theTheme.fontPixelSize)
-        hSp = self.mainConf.pxInt(6)
-
-        # Main Widget Settings
-        self.setContentsMargins(0, 0, 0, 0)
-        self.setAutoFillBackground(True)
-        self.setPalette(self.thePalette)
-
-        buttonStyle = (
-            "QToolButton {{border: none; background: transparent;}} "
-            "QToolButton:hover {{border: none; background: rgba({0},{1},{2},0.2);}}"
-        ).format(*self.theTheme.colText)
-
-        # Show/Hide Details
-        self.showHide = QToolButton(self)
-        self.showHide.setToolButtonStyle(Qt.ToolButtonIconOnly)
-        self.showHide.setStyleSheet(buttonStyle)
-        self.showHide.setIcon(self.theTheme.getIcon("reference"))
-        self.showHide.setIconSize(QSize(fPx, fPx))
-        self.showHide.setFixedSize(QSize(fPx, fPx))
-        self.showHide.clicked.connect(self._doShowHide)
-
-        # Sticky Button
-        stickyOn  = self.theTheme.getPixmap("sticky-on", (fPx, fPx))
-        stickyOff = self.theTheme.getPixmap("sticky-off", (fPx, fPx))
-        stickyIcon = QIcon()
-        stickyIcon.addPixmap(stickyOn, QIcon.Normal, QIcon.On)
-        stickyIcon.addPixmap(stickyOff, QIcon.Normal, QIcon.Off)
-        self.stickyRefs = QToolButton(self)
-        self.stickyRefs.setCheckable(True)
-        self.stickyRefs.setToolButtonStyle(Qt.ToolButtonIconOnly)
-        self.stickyRefs.setStyleSheet(buttonStyle)
-        self.stickyRefs.setIcon(stickyIcon)
-        self.stickyRefs.setIconSize(QSize(fPx, fPx))
-        self.stickyRefs.setFixedSize(QSize(fPx, fPx))
-        self.stickyRefs.toggled.connect(self._doToggleSticky)
-
-        # Labels
-        self.lblRefs = QLabel("References")
-        self.lblRefs.setBuddy(self.showHide)
-        self.lblRefs.setIndent(0)
-        self.lblRefs.setMargin(0)
-        self.lblRefs.setContentsMargins(0, 0, 0, 0)
-        self.lblRefs.setAutoFillBackground(True)
-        self.lblRefs.setFixedHeight(fPx)
-        self.lblRefs.setAlignment(Qt.AlignLeft | Qt.AlignTop)
-        self.lblRefs.setPalette(self.thePalette)
-
-        self.lblSticky = QLabel("Sticky")
-        self.lblSticky.setBuddy(self.stickyRefs)
-        self.lblSticky.setIndent(0)
-        self.lblSticky.setMargin(0)
-        self.lblSticky.setContentsMargins(0, 0, 0, 0)
-        self.lblSticky.setAutoFillBackground(True)
-        self.lblSticky.setFixedHeight(fPx)
-        self.lblSticky.setAlignment(Qt.AlignLeft | Qt.AlignTop)
-        self.lblSticky.setPalette(self.thePalette)
-
-        lblFont = self.font()
-        lblFont.setPointSizeF(0.9*self.theTheme.fontPointSize)
-        self.lblRefs.setFont(lblFont)
-        self.lblSticky.setFont(lblFont)
-
-        # Assemble Layout
-        self.outerBox = QHBoxLayout()
-        self.outerBox.setSpacing(hSp)
-        self.outerBox.addWidget(self.showHide, 0)
-        self.outerBox.addWidget(self.lblRefs, 0)
-        self.outerBox.addWidget(self.stickyRefs, 0)
-        self.outerBox.addWidget(self.lblSticky, 0)
-        self.outerBox.addStretch(1)
-        self.setLayout(self.outerBox)
-
-        logger.debug("GuiDocViewFooter initialisation complete")
-
-        return
-
-    ##
-    #  Slots
-    ##
-
-    def _doShowHide(self):
-        """Toggle the expand/collapse of the panel.
-        """
-        isVisible = self.viewMeta.isVisible()
-        self.viewMeta.setVisible(not isVisible)
-        return
-
-    def _doToggleSticky(self, theState):
-        """Toggle the sticky flag for the reference panel.
-        """
-        logger.verbose("Reference sticky is %s" % str(theState))
-        self.theParent.stickyRef = theState
-        if not theState and self.theParent.theHandle is not None:
-            self.viewMeta.refreshReferences(self.theParent.theHandle)
-        return
-
-# END Class GuiDocViewFooter
-
-class GuiDocViewDetails(QScrollArea):
-
-    def __init__(self, theParent):
-        QScrollArea.__init__(self, theParent)
-
-        logger.debug("Initialising GuiDocViewDetails ...")
-        self.mainConf   = nw.CONFIG
-        self.theParent  = theParent
-        self.theProject = theParent.theProject
-        self.theTheme   = theParent.theTheme
-        self.currHandle = None
-
-        self.refList = QLabel("None")
-        self.refList.setWordWrap(True)
-        self.refList.setAlignment(Qt.AlignTop)
-        self.refList.setScaledContents(True)
-        self.refList.linkActivated.connect(self._linkClicked)
-
-        hCol = self.palette().highlight().color()
-        self.linkStyle = "style='color: rgb({0},{1},{2})'".format(
-            *self.theTheme.colLink
-        )
-
-        # Assemble
-        self.outerWidget = QWidget()
-        self.outerBox = QHBoxLayout()
-        self.outerBox.addWidget(self.refList, 1)
-
-        self.outerWidget.setLayout(self.outerBox)
-        self.setWidget(self.outerWidget)
-
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarAsNeeded)
-        self.setVerticalScrollBarPolicy(Qt.ScrollBarAsNeeded)
-        self.setWidgetResizable(True)
-        self.setMinimumHeight(self.mainConf.pxInt(50))
-
-        logger.debug("GuiDocViewDetails initialisation complete")
-
-        return
-
-    def refreshReferences(self, tHandle):
-        """Update the current list of document references from the
-        project index.
-        """
-        self.currHandle = tHandle
-        if self.theParent.docViewer.stickyRef:
-            return
-
-        theRefs = self.theParent.theIndex.getBackReferenceList(tHandle)
-        theList = []
-        for tHandle in theRefs:
-            tItem = self.theProject.projTree[tHandle]
-            if tItem is not None:
-                theList.append("<a href='#head_%s:%s' %s>%s</a>" % (
-                    tHandle, theRefs[tHandle], self.linkStyle, tItem.itemName
-                ))
-
-        self.refList.setText(", ".join(theList))
-
-        return
-
-    ##
-    #  Internal Functions
-    ##
-
-    def _linkClicked(self, theLink):
-        """Capture the link-click and forward it to the document viewer
-        class for handling.
-        """
-        logger.verbose("Clicked link: '%s'" % theLink)
-        if len(theLink) == 27:
-            tHandle = theLink[6:19]
-            self.theParent.viewDocument(tHandle, theLink)
-        return
-
-# END Class GuiDocViewDetails
